Update transmission logic
This commit is a checkpoint for working code from the test branch.

Rework the transmission logic to just block each packet until it can
send. Each main loop iteration will then send one packet when the bus is
idle.

We also add some code to sync up with the clock through one full cycle.
Not doing this caused some bits to be sent out of sync, which made the
whole packet fail and the keypad wig out. Waiting a full clock cycle
before sending the first bit avoids this problem

Use a queue for the keypresses so when disarming we just add the
keypresses to the queue, then process through them each loop iteration.
Also added a 500ms delay between each transmission to help prevent
clashes, but this can probably be shortened later on once we're 100%
certain the code works properly.

Finally, added some alt keycodes. The keypress type byte in particular
seems to be different to the one from the upstream repo. Possibly
because of the different models, not really sure. But this works for
mine.